move pallet staking in separate file
